refactor: :truck: Move `projects` to `data/` folder
